+++ src/Exceptions/StepTimeoutException.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AlizHarb\FlowForge\Exceptions;
+
+use RuntimeException;
+
+class StepTimeoutException extends RuntimeException
+{
+    public function __construct(string $stepName, int $timeout)
+    {
+        parent::__construct("Step '{$stepName}' timed out after {$timeout} seconds.");
+    }
+}